DEV: Reorganize lib and spec directories
Organize files into logical subdirectories:

lib/mini_tarball/
- headers/: header.rb, header_fields.rb, header_formatter.rb,
            header_writer.rb, user_group_lookup.rb
- validators/: path_validator.rb, source_validator.rb
- writers/: writer.rb, content_writer.rb, null_writer.rb,
            placeholder.rb, placeholder_manager.rb, entry_attributes.rb
- streams/: (unchanged)
- errors.rb, version.rb (top-level)

Specs mirror the lib structure. Integration tests moved to writers/.
