+++ lib/mini_tarball/headers/header.rb
