+++ lib/mini_tarball/headers/header_fields.rb
