+++ lib/mini_tarball/headers/header_formatter.rb
