+++ lib/mini_tarball/headers/header_writer.rb
