+++ lib/mini_tarball/headers/user_group_lookup.rb
