+++ lib/mini_tarball/validators/path_validator.rb
