+++ lib/mini_tarball/validators/source_validator.rb
