+++ lib/mini_tarball/writers/content_writer.rb
