+++ lib/mini_tarball/writers/entry_attributes.rb
