+++ lib/mini_tarball/writers/null_writer.rb
