+++ lib/mini_tarball/writers/placeholder.rb
