+++ lib/mini_tarball/writers/placeholder_manager.rb
