+++ lib/mini_tarball/writers/writer.rb
