+++ spec/integration/writers/gzip_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../integration_helper"
+require_relative "../../integration_helper"
 require "zlib"
 require "tmpdir"
