+++ spec/integration/writers/symlinks_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../integration_helper"
+require_relative "../../integration_helper"
 
 RSpec.describe "Symlinks" do
   it "extracts a symlink with short target" do
