+++ spec/integration/writers/unicode_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../integration_helper"
+require_relative "../../integration_helper"
 
 RSpec.describe "Unicode" do
   it "extracts a file with UTF-8 filename" do
