+++ spec/integration/writers/writer_error_recovery_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../integration_helper"
+require_relative "../../integration_helper"
 
 RSpec.describe "Writer error recovery" do
   let(:default_options) do
