+++ spec/unit/headers/header_formatter_spec.rb
