+++ spec/unit/headers/header_writer_spec.rb
