+++ spec/unit/headers/user_group_lookup_spec.rb
