+++ spec/unit/validators/path_validator_spec.rb
