+++ spec/unit/writers/content_writer_spec.rb
