+++ spec/unit/writers/entry_attributes_spec.rb
