+++ spec/unit/writers/null_writer_spec.rb
