+++ spec/unit/writers/placeholder_manager_spec.rb
