+++ spec/unit/writers/placeholder_spec.rb
