+++ spec/unit/writers/writer_spec.rb
